+++ app/Livewire/Supplier/Index.php
@@ -88,6 +88,7 @@ class Index extends Component
 
     public function delete($id)
     {
+        abort_if(auth()->user()->role === 'admin', 403, 'Admin tidak diizinkan menghapus data.');
         Supplier::find($id)->delete();
     }
 }
